refactor: move GetBasalSeries orchestration from controller into ChartDataService
The controller had 4 extra dependencies (IBasalSeriesBuilder,
ITempBasalRepository, ITherapySettingsResolver, IBasalRateResolver) for
inline orchestration logic. Move that behind IChartDataService so the
controller stays a thin HTTP adapter.
