+++ src/API/Nocturne.API/Controllers/V4/Analytics/ChartDataController.cs
@@ -1,8 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using OpenApi.Remote.Attributes;
 using Nocturne.Core.Contracts.Analytics;
-using Nocturne.Core.Contracts.Profiles.Resolvers;
-using Nocturne.Core.Contracts.V4.Repositories;
 using Nocturne.Core.Models;
 
 namespace Nocturne.API.Controllers.V4.Analytics;
@@ -26,26 +24,14 @@ namespace Nocturne.API.Controllers.V4.Analytics;
 public class ChartDataController : ControllerBase
 {
     private readonly IChartDataService _chartDataService;
-    private readonly IBasalSeriesBuilder _basalSeriesBuilder;
-    private readonly ITempBasalRepository _tempBasalRepository;
-    private readonly ITherapySettingsResolver _therapySettingsResolver;
-    private readonly IBasalRateResolver _basalRateResolver;
     private readonly ILogger<ChartDataController> _logger;
 
     public ChartDataController(
         IChartDataService chartDataService,
-        IBasalSeriesBuilder basalSeriesBuilder,
-        ITempBasalRepository tempBasalRepository,
-        ITherapySettingsResolver therapySettingsResolver,
-        IBasalRateResolver basalRateResolver,
         ILogger<ChartDataController> logger
     )
     {
         _chartDataService = chartDataService;
-        _basalSeriesBuilder = basalSeriesBuilder;
-        _tempBasalRepository = tempBasalRepository;
-        _therapySettingsResolver = therapySettingsResolver;
-        _basalRateResolver = basalRateResolver;
         _logger = logger;
     }
 
@@ -125,29 +111,7 @@ public class ChartDataController : ControllerBase
             if (endTime <= startTime)
                 return Problem(detail: "endTime must be greater than startTime", statusCode: 400, title: "Bad Request");
 
-            var defaultBasalRate = 1.0;
-            var hasData = await _therapySettingsResolver.HasDataAsync(cancellationToken);
-            if (hasData)
-                defaultBasalRate = await _basalRateResolver.GetBasalRateAsync(endTime, ct: cancellationToken);
-
-            var tempBasals = (await _tempBasalRepository.GetAsync(
-                from: DateTimeOffset.FromUnixTimeMilliseconds(startTime).UtcDateTime,
-                to: DateTimeOffset.FromUnixTimeMilliseconds(endTime).UtcDateTime,
-                device: null,
-                source: null,
-                limit: 131072,
-                offset: 0,
-                descending: false,
-                ct: cancellationToken
-            )).ToList();
-
-            var basalSeries = await _basalSeriesBuilder.BuildAsync(
-                tempBasals,
-                startTime,
-                endTime,
-                defaultBasalRate,
-                cancellationToken
-            );
+            var basalSeries = await _chartDataService.GetBasalSeriesAsync(startTime, endTime, cancellationToken);
 
             return Ok(basalSeries);
         }
